+++ pkgs/checks/lib/src/extensions/async.dart
@@ -21,19 +21,24 @@ extension FutureChecks<T> on Subject<Future<T>> {
   /// and [completionCondition] has optionally been checked.
   @awaitNotRequired
   Future<void> completes([AsyncCondition<T>? completionCondition]) {
-    return context.nestAsync<T>(() => ['completes to a value'], (actual) async {
-      try {
-        return Extracted.value(await actual);
-      } catch (e, st) {
-        return Extracted.rejection(
-          actual: ['a future that completes as an error'],
-          which: [
-            ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
-            ...indent(LineSplitter.split(st.toString())),
-          ],
-        );
-      }
-    }, completionCondition);
+    return context.nestAsync<T>(
+      () => ['completes to a value'],
+      addPredicate: (predicateNoun) => 'completes to $predicateNoun',
+      (actual) async {
+        try {
+          return Extracted.value(await actual);
+        } catch (e, st) {
+          return Extracted.rejection(
+            actual: ['a future that completes as an error'],
+            which: [
+              ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
+              ...indent(LineSplitter.split(st.toString())),
+            ],
+          );
+        }
+      },
+      completionCondition,
+    );
   }
 
   /// Expects that the `Future` never completes as a value or an error.
@@ -85,6 +90,7 @@ extension FutureChecks<T> on Subject<Future<T>> {
   Future<void> throws<E extends Object>([AsyncCondition<E>? errorCondition]) {
     return context.nestAsync<E>(
       () => ['completes to an error${E == Object ? '' : ' of type $E'}'],
+      addPredicate: (predicateNoun) => 'throws $predicateNoun',
       (actual) async {
         try {
           return Extracted.rejection(
@@ -148,26 +154,31 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
   /// ended, and the [emittedCondition] has optionally been checked.
   @awaitNotRequired
   Future<void> emits([AsyncCondition<T>? emittedCondition]) {
-    return context.nestAsync<T>(() => ['emits a value'], (actual) async {
-      if (!await actual.hasNext) {
-        return Extracted.rejection(
-          actual: ['a stream'],
-          which: ['closed without emitting enough values'],
-        );
-      }
-      try {
-        await actual.peek;
-        return Extracted.value(await actual.next);
-      } catch (e, st) {
-        return Extracted.rejection(
-          actual: prefixFirst('a stream with error ', literal(e)),
-          which: [
-            'emitted an error instead of a value at:',
-            ...indent(LineSplitter.split(st.toString())),
-          ],
-        );
-      }
-    }, emittedCondition);
+    return context.nestAsync<T>(
+      () => ['emits a value'],
+      addPredicate: (predicateNoun) => 'emits $predicateNoun',
+      (actual) async {
+        if (!await actual.hasNext) {
+          return Extracted.rejection(
+            actual: ['a stream'],
+            which: ['closed without emitting enough values'],
+          );
+        }
+        try {
+          await actual.peek;
+          return Extracted.value(await actual.next);
+        } catch (e, st) {
+          return Extracted.rejection(
+            actual: prefixFirst('a stream with error ', literal(e)),
+            which: [
+              'emitted an error instead of a value at:',
+              ...indent(LineSplitter.split(st.toString())),
+            ],
+          );
+        }
+      },
+      emittedCondition,
+    );
   }
 
   /// Expects that the stream emits an error of type [E].
@@ -191,6 +202,7 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
   ]) {
     return context.nestAsync<E>(
       () => ['emits an error${E == Object ? '' : ' of type $E'}'],
+      addPredicate: (predicateNoun) => 'emits error $predicateNoun',
       (actual) async {
         if (!await actual.hasNext) {
           return Extracted.rejection(
@@ -297,7 +309,15 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
                 if (satisfiedCount > 0)
                   'satisfied $satisfiedCount conditions then',
                 'failed to satisfy the condition at index $satisfiedCount',
-                if (failure.detail.depth > 0) ...[
+                if (failure.detail.collapsedDetails case (
+                  final expected,
+                  final actual,
+                )) ...[
+                  'because it:',
+                  '  Expected: $expected',
+                  '  Actual: $actual',
+                  if (which != null) ...indent(prefixFirst('Which: ', which)),
+                ] else if (failure.detail.depth > 0) ...[
                   'because it:',
                   ...indent(
                     failure.detail.actual.skip(1),
@@ -378,7 +398,14 @@ extension StreamChecks<T> on Subject<StreamQueue<T>> {
           final which = failure.rejection.which;
           final detail = failure.detail;
           final failed = 'failed the condition at index $index';
-          if (detail.depth > 0) {
+          if (detail.collapsedDetails case (final expected, final actual)) {
+            return [
+              '$failed because it:',
+              '  Expected: $expected',
+              '  Actual: $actual',
+              if (which != null) ...indent(prefixFirst('Which: ', which)),
+            ];
+          } else if (detail.depth > 0) {
             return [
               '$failed because it:',
               ...indent(detail.actual.skip(1), detail.depth - 1),